moved cache stats collection to a seperate file and removed redundant configurations
